pyproject.toml
[project]
name = "mcp-sharepoint"
version = "0.1.2"
description = "MCP Server to integrate with SharePoint"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
 "mcp>=1.2.1",
 "msal>=1.31.0",
 "msal-extensions>=1.1.0",
 "python-dotenv>=1.0.0",
 "requests>=2.32.0",
]

[[project.authors]]
name = "sofias tech"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.scripts]
mcp-sharepoint = "mcp_sharepoint:main"